Previous commit asked about semantics.
I'm still not sure if `GetInvokerType()` should *always* return a
`Type`, or only return a type that is an invoker type.

Leaving that aside, there is a question around parameter validation:
there wasn't any before.

A common (if inconsistently applied) pattern in java-interop is to
have a public non-virtual method that does parameter validation, and
then calls a protected virtual method that does the actual work.
Apply that pattern here, calling the protected method
`GetInvokerTypeCore()`.