Copy JCW `.jar` to OutputPath
Context: 37f68a3895de221ed1f26d02b4aba5edfa36282a

Commit 37f68a38 contained a comment+TODO:

	# Something is missing in the targets that `Hello-Java.Base.jar` isn't copied; copy it.
	# TODO: fix this.
	cp samples/Hello-Java.Base/bin/Release/Hello-Java.Base.jar samples/Hello-Java.Base/bin/Release/osx-x64/publish

Fix that.  The extra `cp` is no longer needed.

Fixing this required three things:

 1. Adding the JCW `.jar` to `@(Content)` (or `@(None)`, or…)
 2. Setting `%(Content.CopyToOutputDirectory)`=PreserveNewest
 3. Setting `%(Content.TargetPath)`

`%(TargetPath)` is used to control the filename within the final
`$(OutputPath)` directory tree; *without* `%(TargetPath)`, then
`%(Identity)` is used, and since `%(Identity)` is now within
`$(IntermediateOutputPath)`, that's *wrong*.

Rename `$(JavaOutputJar)` to `$(JavaOutputJarName)`, as the
new implied semantic is that it *cannot* contain a directory; it's
*just* the filename.

Update the `_JavaCreateJcws` so that instead of being a post-Build
target it's instead a post-CoreCompile target.  This is required
because post-Build is too late; MSBuild tries to copy the `@(Content)`
entries to `$(OutputPath)` *before the end* of `Build`, and as a
post-Build step it didn't have a chance to *create* the jar yet!
A post-CoreCompile target works nicer in this context.

Because we're now post-CoreCompile, `$(TargetPath)` doesn't exist.
(That's created during the build, by copying the intermediate
assembly to the outptu path.)  Update `_JavaCreateJcws` to use
`@(IntermediateAssembly)` instead of `$(TargetPath)`.

Update unit tests accordingly.