Standardize `$(LangVersion)` in `netstandard2.0` projects (#1322)
The default `$(LangVersion)` for `netstandard2.0` projects is
[C# 7.3][0], which does not contain many of the niceties we would
like to use.  Throughout our history we have updated individual
projects to use the latest language features, but this has led to
inconsistencies.

Instead, standardize this in `Directory.Build.targets` to use a
common C# language version for `netstandard2.0` projects.

Note that we standardized on [C# 13.0][1] -- the default in .NET 9 --
instead of `latest`, as `latest` is discouraged because it can change
based on what versions are installed on a machine.

[0]: https://learn.microsoft.com/en-us/dotnet/csharp/whats-new/csharp-version-history#c-version-73
[1]: https://learn.microsoft.com/en-us/dotnet/csharp/whats-new/csharp-version-history#c-version-13