﻿[JavaCallableWrappers] Add XML import/export. (#1317)
Context: https://github.com/dotnet/android/pull/9893

Create XML import/export adapters for Java Callable Wrappers (JCWs).
This allows us to serialize the JCWs we find in the new
`FindJavaObjectsStep` linker step (dotnet/android#9893) to disk that,
and later use that serialized data in `_GenerateJavaStubs`.

This file also allows us to cache the assembly scanning to disk so we
don't need to rerun it on incremental builds.

This format is likely a "work in progress".  As we move other
`Java.Lang.Object`-related scanning to linker steps we will likely find
additional data they require that needs to go into this file, or that
this format isn't even suitable and we use something else instead.

Sample JCW XML content:

	<types was_scanned="true">
	  <type
	      name="MainActivity"
	      package="crc645107ba1b8b6ee4d3"
	      application_java_class="android.app.Application"
	      mono_runtime_initialization="mono.MonoPackageManager.LoadApplication (context);"
	      extends_type="android.app.Activity"
	      partial_assembly_qualified_name="tempbuild.MainActivity, tempbuild">
	    <constructors>
	      <constructor
	        name="MainActivity"
	        method="n_.ctor:()V:"
	        jni_signature="()V"
	        managed_parameters=""
	        retval="void"
	        is_dynamically_registered="True" />
	    </constructors>
	    <methods>
	      <method
	        name="clone"
	        method="n_clone:()Ljava/lang/Object;:GetCloneHandler"
	        jni_signature="()Ljava/lang/Object;"
	        retval="java.lang.Object" />
	      <method
	        name="onCreate"
	        method="n_onCreate:(Landroid/os/Bundle;)V:GetOnCreate_Landroid_os_Bundle_Handler"
	        jni_signature="(Landroid/os/Bundle;)V"
	        params="android.os.Bundle p0"
	        retval="void"
	        super_call="p0"
	        activate_call="p0" />
	    </methods>
	  </type>
	</types>

Most XML attributes correspond to members on `CallableWrapperType`,
`CallableWrapperMethod`, etc.

The XML file is a *per-assembly* data store; `//types/@was_scanned`
specifies whether or not the assembly was scanned for JCWs.
If the assembly did not contain JCWs (e.g.
`System.Private.CoreLib.dll`), then `was_scanned` will be False.