[build] Change `$(JavacClasspathSeparator)` to `$(JavaPathSeparator)` (#1373)
In `build-tools/Java.Interop.Sdk/Sdk/Sdk.props`, the property
`$(JavaPathSeparator)` is defined, but the condition for the
non-Windows case references an undeclared property
`$(JavacClasspathSeparator)`.

It causes on Windows, javac -classpath with multiple jars fails.

This is likely a typo. This change corrects it to
`$(JavaPathSeparator)` to match the property being defined.